refactor(data_bus): replace integer return types with uint32_t in channel enqueue functions
- Updated the return type of channel_enqueue_block and related functions from int to uint32_t for better type consistency and clarity.
- Replaced boolean active checks with atomic operations in data_bus_channel and data_bus_consumer to enhance thread safety.
- Improved logging to cast atomic values to int for clearer debug output during dispatch operations.

+++ src/data_bus/data_bus_consumer.c
@@ -206,7 +206,7 @@ void data_bus_consumer_dispatch(data_bus_channel_t *ch, data_bus_block_t *block)
 	 * 每个消费者获得一份隐式引用。
 	 */
 	LOG_DBG("Dispatch ch='%s' seq=%u ref+%u=%d", ch->name, block->seq,
-		active_count, atomic_get(&block->ref_count));
+		active_count, (int)atomic_get(&block->ref_count));
 
 	atomic_add(&block->ref_count, active_count);
